Changed Ids from Int to Long

+++ app/src/main/java/com/towich/cosmicintrigue/ui/adapters/VoteAdapter.kt
@@ -9,9 +9,9 @@ import com.towich.cosmicintrigue.R
 import com.towich.cosmicintrigue.databinding.VoteViewBinding
 import com.towich.cosmicintrigue.data.model.Player
 
-class VoteAdapter(var onVote:(Int?)->Unit, players: List<Player>, val userId : Int) :RecyclerView.Adapter<VoteAdapter.VoteViewHolder>(){
+class VoteAdapter(var onVote:(Long?)->Unit, players: List<Player>, val userId : Long) :RecyclerView.Adapter<VoteAdapter.VoteViewHolder>(){
     class VoteViewHolder(val binding: VoteViewBinding):RecyclerView.ViewHolder(binding.root)
-    class AdaptedPlayer(var id:Int?, var login:String, var isUser:Boolean, var isVoted:Boolean, var numVotes: Int)
+    class AdaptedPlayer(var id:Long?, var login:String, var isUser:Boolean, var isVoted:Boolean, var numVotes: Int)
 
 
     var votedId: Int = 0
@@ -23,8 +23,8 @@ class VoteAdapter(var onVote:(Int?)->Unit, players: List<Player>, val userId : I
         }
     }
 
-    public fun endVote(Votes: List<Pair<Int?,Int>>){
-        val (listA:List<Int?>, listB:List<Int>) =Votes.unzip()
+    public fun endVote(Votes: List<Pair<Long?,Int>>){
+        val (listA:List<Long?>, listB:List<Int>) =Votes.unzip()
         var y = listA.get(2)
         for(i in Users.indices)
         {
